requirements.txt
pyfrotz
google-adk
python-dotenv
TTS
numpy
torch
pygame
google-cloud-aiplatform>=1.38.1